+++ examples/phone_number/with-search-criteria/outputs.tf
@@ -1,0 +1,4 @@
+output "phone_number" {
+  description = "The Purchased Phone Number"
+  value       = twilio_phone_number.phone_number
+}