+++ examples/phone_number/with-search-criteria/versions.tf
@@ -1,0 +1,11 @@
+terraform {
+  required_version = ">= 0.13"
+
+  required_providers {
+    twilio = {
+      source  = "RJPearson94.com/twilio/twilio"
+      version = ">= 0.1.0"
+    }
+  }
+}
+